--- STIW3054 Project/CountCharactersInterface.java
@@ -1,5 +1,0 @@
-import java.util.concurrent.atomic.AtomicInteger;
-
-public interface CountCharactersInterface {
-    public AtomicInteger calculateTotalCharacters();
-}
