--- STIW3054 Project/CountCommonsMathInterface.java
@@ -1,5 +1,0 @@
-public interface CountCommonsMathInterface {
-    public double countMean();
-    public double countVariance();
-    public double countSD();
-}
