--- STIW3054 Project/CountWordsInterface.java
@@ -1,5 +1,0 @@
-import java.util.concurrent.atomic.AtomicInteger;
-
-public interface CountWordsInterface {
-    public AtomicInteger calculateTotalWords();
-}
