+++ STIW3054_Project/src/main/java/CountCharactersInterface.java
@@ -1,0 +1,5 @@
+import java.util.concurrent.atomic.AtomicInteger;
+
+public interface CountCharactersInterface {
+    public AtomicInteger calculateTotalCharacters();
+}
