+++ STIW3054_Project/src/main/java/CountCommonsMathInterface.java
@@ -1,0 +1,5 @@
+public interface CountCommonsMathInterface {
+    public double countMean();
+    public double countVariance();
+    public double countSD();
+}
