+++ STIW3054_Project/src/main/java/CountEachCharacters.java
