+++ STIW3054_Project/src/main/java/CountEachCharactersInterface.java
