+++ STIW3054_Project/src/main/java/CountWordsInterface.java
@@ -1,0 +1,5 @@
+import java.util.concurrent.atomic.AtomicInteger;
+
+public interface CountWordsInterface {
+    public AtomicInteger calculateTotalWords();
+}
