+++ STIW3054_Project/src/main/java/GraphNormalization.java
