+++ STIW3054_Project/src/main/java/GraphNormalizationInterface.java
